# src/main/java/com/openbanking/scheduled/payments/model/OBReadScheduledPayment3.java
package com.openbanking.scheduled.payments.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadScheduledPayment3
 */
@Getter
@Setter
public class OBReadScheduledPayment3 implements Serializable {
	private static final long serialVersionUID = 1L;

}